Move cells into their own files
